cast the string size to gulong and use %lu for printing to avoid 64bit
Sat Sep 29 02:33:14 2001  George Lebl <jirka@5z.com>

	* tests/testglib.c (main): cast the string size to gulong and use %lu
	  for printing to avoid 64bit issues

	* glib/gstrfuncs.[ch] (g_ascii_strncasecmp): Change the 'n' argument
	  to gsize instead guint.

+++ glib/gstrfuncs.h
@@ -129,7 +129,7 @@ gint                  g_ascii_strcasecmp  (const gchar *s1,
 					   const gchar *s2);
 gint                  g_ascii_strncasecmp (const gchar *s1,
 					   const gchar *s2,
-					   guint        n);
+					   gsize        n);
 gchar*                g_ascii_strdown     (const gchar *str,
 					   gssize       len);
 gchar*                g_ascii_strup       (const gchar *str,
